Support custom id for footbibliography.
