+++ src/sphinxcontrib/bibtex/foot_directives.py
@@ -7,6 +7,8 @@
 from typing import TYPE_CHECKING, cast
 from docutils.parsers.rst import Directive
 
+from .bibfile import _make_ids
+
 if TYPE_CHECKING:
     from sphinx.environment import BuildEnvironment
     from .domain import BibtexDomain
@@ -28,6 +30,9 @@ class FootBibliographyDirective(Directive):
         env = cast("BuildEnvironment", self.state.document.settings.env)
         foot_old_refs = env.temp_data.setdefault("bibtex_foot_old_refs", set())
         foot_new_refs = env.temp_data.setdefault("bibtex_foot_new_refs", set())
+        footbibliography_count = \
+            env.temp_data["bibtex_footbibliography_count"] = \
+            env.temp_data.get("bibtex_footbibliography_count", 0) + 1
         if not foot_new_refs:
             return []
         else:
@@ -41,4 +46,11 @@ class FootBibliographyDirective(Directive):
             domain = cast("BibtexDomain", env.get_domain('cite'))
             for bibfile in domain.bibdata.bibfiles:
                 env.note_dependency(bibfile)
+            foot_bibliography['ids'] += _make_ids(
+                docname=env.docname, lineno=self.lineno,
+                ids=set(self.state.document.ids.keys()),
+                raw_id=env.app.config.bibtex_footbibliography_id.format(
+                    footbibliography_count=footbibliography_count))
+            self.state.document.note_explicit_target(
+                foot_bibliography, foot_bibliography)
             return [foot_bibliography]
