+++ test/test_bibliography.py
@@ -231,8 +231,8 @@ def test_bibliography_custom_ids(app, warning) -> None:
     output = (app.outdir / "index.html").read_text(encoding='utf-8')
     assert '<p id="bibliography-id-1">' in output
     assert '<p id="bibliography-id-2">' in output
-    #assert '<p id="footbibliography-id-1">' in output
-    #assert '<p id="footbibliography-id-2">' in output
+    assert '<p id="footbibliography-id-1">' in output
+    assert '<p id="footbibliography-id-2">' in output
     match1 = html_citations(text='.*Evensen.*').search(output)
     match2 = html_citations(text='.*Mandel.*').search(output)
     match3 = html_citations(text='.*Lorenc.*').search(output)
